# project/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SMM Project API",
        "description": "API for SMM project management system",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/v1",
            "description": "Development server"
        }
    ],
    "paths": {
        "/activities": {
            "get": {
                "tags": [
                    "Activities"
                ],
                "summary": "Get pending activities",
                "description": "Retrieve list of pending activities for authenticated user",
                "operationId": "0912b46e1df7261b928de0779141465d",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Activity"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/activities/status-batch": {
            "patch": {
                "tags": [
                    "Activities"
                ],
                "summary": "Update multiple activity statuses",
                "description": "Update status for multiple activities in a single request",
                "operationId": "cedea0d96c868966b074b66fde1a3f51",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ids",
                                    "status"
                                ],
                                "properties": {
                                    "ids": {
                                        "description": "Array of activity IDs to update",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3
                                        ]
                                    },
                                    "status": {
                                        "description": "New status for activities",
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "paid",
                                            "read",
                                            "declined"
                                        ],
                                        "example": "read"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Statuses updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Statuses updated successfully"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "read"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Activity": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "type": "string",
                        "example": "task_created"
                    },
                    "description": {
                        "type": "string",
                        "example": "New task was created"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "read",
                            "archived"
                        ],
                        "example": "pending"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-21T10:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}